[Rebase] Merge IPEX frontend code for 2.3 release (#4017)
* modify frontend.py

* sync logger

* change warning api

* merge cpu/launch/*

* merge cpu/nn/*

* merge cpu/runtime/*

* merge cpu/tpp/*

* merge cpu/*

* merge fx/*

* merge nn/functional/*
merge nn/modules/*

* merge nn/modules/*
merge utils/weight_only_quantization.py

* fix import
format

* merge

* remove IPEX_TILE_AS_DEVICE warning
version to 2.3.10

* merge quantization/**

* fix import error

* merge nn/*

* port quantization/fp8
fix import

* merge transformers/*

* merge cpu/autocast

* merge cpu

* sync cpu transformers/models/cpu/*
modify transforemrs/optimize

* merge mha_fusion.py

* merge llm

* merge _inductor

* import quantization/_GPTQ

* fix utils path

* align cpu gpu fp8 schema

* Revert "align cpu gpu fp8 schema"

This reverts commit c87034ee15a55210d78cdea50fd1ea9b17ced5ce.

* align gpu fp8 api to cpu

* import _inductor.cpu

* fix ut

* fix test_fast_layer_norm

* fix FastLayerNorm

* fix test_rms_norm

* format

* merge optim diff

* cherry pick cpu #2741

* Add cpu supported models check. (#4119)

* fix test_softmax

* Add cpu supported models check in optimize.py. (#4121)

* Add cpu supported models check.

* Fix.

---------


* merge cpu ut diff

* format
fix cpu import

* fix import

* fix tests/gpu/examples/test_sdp_bias_bc.py

---------

+++ tests/gpu/examples/test_sdp_bias_bc.py
@@ -25,11 +25,11 @@ class TestTorchMethod(TestCase):
             ):
                 print("seq_lenth is ", seq_lenth)
                 print("bias_size", bias_size)
-                query_states = torch.rand((b, n, q_len, attn_head_size))
-                key_states = torch.rand((b, n, kv_len, attn_head_size))
-                value_states = torch.rand((b, n, kv_len, attn_head_size))
-                grad = torch.rand((b, n, q_len, attn_head_size))
-                bias = torch.rand(bias_size)
+                query_states = torch.rand((b, n, q_len, attn_head_size), dtype=dtype)
+                key_states = torch.rand((b, n, kv_len, attn_head_size), dtype=dtype)
+                value_states = torch.rand((b, n, kv_len, attn_head_size), dtype=dtype)
+                grad = torch.rand((b, n, q_len, attn_head_size), dtype=dtype)
+                bias = torch.rand(bias_size, dtype=dtype)
 
                 query_states_xpu = query_states.to(dtype).xpu()
                 key_states_xpu = key_states.to(dtype).xpu()
@@ -56,25 +56,25 @@ class TestTorchMethod(TestCase):
                 r_cpu.backward(grad)
                 r_xpu.backward(grad_xpu)
 
-                self.assertEqual(r_cpu, r_xpu.cpu().float(), atol=1e-2, rtol=1e-1)
+                self.assertEqual(r_cpu.float(), r_xpu.cpu().float(), atol=1e-2, rtol=1e-1)
                 self.assertEqual(
-                    query_states.grad,
+                    query_states.grad.float(),
                     query_states_xpu.grad.cpu().float(),
                     atol=1e-2,
                     rtol=1e-1,
                 )
                 self.assertEqual(
-                    key_states.grad,
+                    key_states.grad.float(),
                     key_states_xpu.grad.cpu().float(),
                     atol=1e-2,
                     rtol=1e-1,
                 )
                 self.assertEqual(
-                    value_states.grad,
+                    value_states.grad.float(),
                     value_states_xpu.grad.cpu().float(),
                     atol=1e-2,
                     rtol=1e-1,
                 )
                 self.assertEqual(
-                    bias.grad, bias_xpu.grad.cpu().float(), atol=1e-2, rtol=1e-1
+                    bias.grad.float(), bias_xpu.grad.cpu().float(), atol=1e-2, rtol=1e-1
                 )
